requirements.txt
Flask
Flask-Restful
Flask-JWT
Flask-SQLAlchemy
uwsgi
psycopg2
pyxero